open-source-references/AI-Safety-Compliance-Officer/requirements.txt
# Computer Vision
ultralytics>=8.0.0
opencv-python>=4.8.0
onnx>=1.15.0
onnxruntime>=1.16.0

# AI Agent
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
langsmith>=0.1.0
openai>=1.0.0

# PDF Generation
reportlab>=4.0.0
Pillow>=10.0.0

# Email
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# AWS SDK (for production deployment)
boto3>=1.28.0
botocore>=1.31.0

# Web Interface (Optional)
flask>=3.0.0
flask-cors>=4.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
